checking role

+++ stockTable/Controllers/EquipmentController.cs
@@ -126,50 +126,64 @@ namespace stockTable.Controllers
         [HttpGet]
         public async Task<IActionResult> Edit(int id)
         {
-            var equipment = await _equipmentRepository.GetById(id);
-            var statuses = await _statusRepository.GetAll();
-            if (equipment == null) return View("Error");
-            var vModel = new EditEqViewModel()
+            if(User.IsInRole("admin") || User.IsInRole("editor"))
             {
-                Equipment = equipment,
-                Statuses = statuses,
-                Document = equipment.Document
-            };
+                var equipment = await _equipmentRepository.GetById(id);
+                var statuses = await _statusRepository.GetAll();
+                if (equipment == null) return View("Error");
+                var vModel = new EditEqViewModel()
+                {
+                    Equipment = equipment,
+                    Statuses = statuses,
+                    Document = equipment.Document
+                };
 
-            return View(vModel);
+                return View(vModel);
+            }
+            return RedirectToAction("Index", "Equipment");
+                
         }
 
         [HttpPost]
         public async Task<IActionResult> Edit(int id, EditEqViewModel model)
         {
-            if(!ModelState.IsValid)
+            if(User.IsInRole("admin") || User.IsInRole("editor"))
             {
-                ModelState.AddModelError("", "Failed to edit Equipment");
-                model.Statuses = await _statusRepository.GetAll();
-                return View(model);
+                if (!ModelState.IsValid)
+                {
+                    ModelState.AddModelError("", "Failed to edit Equipment");
+                    model.Statuses = await _statusRepository.GetAll();
+                    return View(model);
+                }
+
+                var document = model.Document;
+                var equipment = model.Equipment;
+                equipment.Document = document;
+                _documentRepository.Update(document);
+                _equipmentRepository.Update(equipment);
             }
-            
-            var document = model.Document;
-            var equipment = model.Equipment;
-            equipment.Document = document;
-            _documentRepository.Update(document);
-            _equipmentRepository.Update(equipment);
-            return RedirectToAction("Index","Equipment");
+            return RedirectToAction("Index", "Equipment");
+
         }
 
         public async Task<IActionResult> Delete(int id)
         {
-            var equipment = await _equipmentRepository.GetById(id);
-            var document = await _documentRepository.GetById(equipment.IdDocument);
-            if ( equipment != null)
+            if (User.IsInRole("admin") || User.IsInRole("editor"))
             {
-                
-                _equipmentRepository.Delete(equipment);
-                _documentRepository.Delete(document!);
-                return RedirectToAction("Index", "Equipment");
+                var equipment = await _equipmentRepository.GetById(id);
+                var document = await _documentRepository.GetById(equipment.IdDocument);
+                if (equipment != null)
+                {
+
+                    _equipmentRepository.Delete(equipment);
+                    _documentRepository.Delete(document!);
+                    return RedirectToAction("Index", "Equipment");
+                }
+                else
+                    return View("Error");
             }
-            else
-                return View("Error");
+            return RedirectToAction("Index", "Equipment");
+
 
         }
     }
